+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v1.0.43
 * Fixed crash L2Hostility, which tried to access level during virtual chunk <init>.
 * Fixed crash with jGUI, they don't use mixin correctly
+* Avoid using stream api for LinkedHashMap